added an AstExpressionStatement that allows you to skip the 'let _ = ' part when calling functions with side effects, fixed a bug in the compiler where it would crash if it couldnt remove a build directory (oops)

+++ lcaml_py/compyla/__main__.py
@@ -124,7 +124,8 @@ if __name__ == "__main__":
     if not os.path.exists(build_dir):
         os.mkdir(build_dir)
     build_dir = os.path.join(build_dir, "compiled_lcaml")
-    shutil.rmtree(build_dir)
+    if os.path.exists(build_dir):
+        shutil.rmtree(build_dir)
     os.mkdir(build_dir)
 
     for file in args.files:
